+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ## [unreleased]
 
+- Update dependencies including React 19.0.0
+
 ### Added
 
 - `sandstone/Icon` supported icon list, adding new icon `calibration`
@@ -17,8 +19,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Fixed
 
+- `sandstone/ContextualPopupDecorator` to update popup position properly when the DOM tree changes
 - `sandstone/Scroller` to focus properly when the spottable node is bigger than the size of viewport by voice control
 - `sandstone/Scroller` to prevent the native scrolling behavior caused by keydown events when a popup is open
+- `sandstone/TooltipDecorator` to hide a tooltip when tapping outside of disabled component
 
 ## [3.0.0-alpha.3] - 2024-12-02